+++ regex/src/com.oracle.truffle.regex/src/com/oracle/truffle/regex/flavor/oracledb/OracleDBRegexParser.java
@@ -48,6 +48,7 @@ import com.oracle.truffle.regex.RegexLanguage;
 import com.oracle.truffle.regex.RegexSource;
 import com.oracle.truffle.regex.RegexSyntaxException;
 import com.oracle.truffle.regex.RegexSyntaxException.ErrorCode;
+import com.oracle.truffle.regex.UnsupportedRegexException;
 import com.oracle.truffle.regex.charset.ClassSetContents;
 import com.oracle.truffle.regex.charset.CodePointSet;
 import com.oracle.truffle.regex.charset.CodePointSetAccumulator;
@@ -169,7 +170,7 @@ public final class OracleDBRegexParser implements RegexParser {
                     break;
                 case quantifier:
                     if (prevKind == Token.Kind.quantifier) {
-                        throw syntaxError(OracleDBErrorMessages.NESTED_QUANTIFIER, ErrorCode.InvalidQuantifier);
+                        throw new UnsupportedRegexException(OracleDBErrorMessages.NESTED_QUANTIFIER);
                     }
                     if (astBuilder.getCurTerm() == null || prevKind == Token.Kind.captureGroupBegin) {
                         // quantifiers without target are ignored